수업등록 및 자유게시판 추가

+++ src/main/java/com/example/idollbom/domain/dto/boarddto/CommunityDTO.java
@@ -15,6 +15,6 @@ public class CommunityDTO {
     private int parentPostViews;
     private LocalDateTime parentPostRegisterDate;
     private LocalDateTime parentPostUpdateDate;
-    private int parentNumber;
+    private Long parentNumber;
 
 }
